+++ src/onegov.wtfs/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds user import script.
+  [href]
+
 1.3.2 (2019-06-18)
 ~~~~~~~~~~~~~~~~~~~